Moving the wrapping to main Mapdl, being common to gpc, console and corba.
