+++ ansys/mapdl/core/mapdl_grpc.py
@@ -64,7 +64,6 @@ from ansys.mapdl.core.common_grpc import (
 )
 from ansys.mapdl.core import __version__, _LOCAL_PORTS
 from ansys.mapdl.core import check_version
-from ansys.mapdl.core.commands import CommandListingOutput, CMD_LISTING
 
 
 TMP_VAR = '__tmpvar__'
@@ -312,18 +311,6 @@ class MapdlGrpc(_MapdlCore):
         # only cache process IDs if launched locally
         if self._local and "exec_file" in kwargs:
             self._cache_pids()
-
-        # Wrapping LISTING FUNCTIONS.
-        def wrap_listing_function(func):
-            @wraps(func)
-            def inner_wrapper(*args, **kwargs):
-                return CommandListingOutput(func(*args, **kwargs))
-            return inner_wrapper
-
-        for name in dir(self):
-            if name[0:4].upper() in CMD_LISTING:
-                func = self.__getattribute__(name)
-                setattr(self, name, wrap_listing_function(func))
 
     def _verify_local(self):
         """Check if Python is local to the MAPDL instance."""
